requirements.txt
amqp==5.1.1
asgiref==3.6.0
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.19
djangorestframework==3.14.0
importlib-metadata==6.6.0
kombu==5.2.4
prompt-toolkit==3.0.38
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.5.0
vine==5.0.0
wcwidth==0.2.6
zipp==3.15.0
